1、增加了关于Float的工具类 2、准备重新计算缩放算法 3、修复了一些bug

+++ chartviewlib/src/main/java/com/example/will/chartviewlib/ChartInfo/BackgroundInfo/IScaleInfo.java
@@ -74,5 +74,5 @@ public interface IScaleInfo {
      * @param min
      * @param max
      */
-    void setYRange(int min, int max);
+    void setYRange(float min, float max);
 }
